fix: move auth file into auth folder
